Update PLDForm.py
I didn't fully test everything but I added functionality for xor to separate terms as well as parentheses to control the order in which the query is evaluated. I also fixed a sorting error that prevented the metadata for records with more than 10 targets from matching up with the labelled target.